Fix login action in user store (#1328)
* Fix login action in user store

Unlike the `verifyLogin` action in the user store, the `login` action does
not add the logged in user to the users store. This causes an error on
pages that do not explicitly load users, such as on the manage
assignment page, where the submission uploader shows the name of the
logged in user by default.

* Stop loading snippets in the login actions

We load the snippets when we need them already.