Fix a race condition when creating a AutoTest run by submitting work
We used to kick of two tasks when somebody handed in a new submission, 1) a task
to update the amount of runners needed for the run, and 2) a task to update the
`results` metadata in the broker (that is used by the broker for health
checks). However, the task to update the metadata passes `error_on_create` as
`True` to the broker (as it shouldn't be creating tasks, it should only be
updating existing tasks and augmenting their metadata).

This however creates a race condition: if the second task (updating the results
metadata) starts before the first one (updating the amount of runners and if
needed creating the run at the broker) and the broker didn't know of this run,
the second task fails (however later retries will work).

This race condition is fixed in this commit in two ways:
1) The task to register a job in the broker now simply also send the `results`
metadata to the broker.
2) The task to update the amount of needed runners now has an extra keyword
parameter (`always_update_latest_results`) and if that is true the task will
always update the `results` metadata, even if the amount of runners registered
at the broker is correct.