Improve handling for failing lxc starts or shutdowns (#1325)
It sometimes happens that a container fails to start or shutdown, this PR improves the handling for these cases. If a container fails to start it seems that it wil never start again on that runner, so we now throw a subclass of `StopRunningTestsException` to kill the runner and reset all active tests to not started again. For shutdowns it is even less clear why they fail, however this PR start retrying shutdowns a few times before bailing and throwing a subclass of `cg_worker_pool.KillWorkerException` (which we now also handle correctly by resetting the result back to not started).

Co-authored-by: Olmo Kramer <olmo@codegrade.com>