Fix infinite loaders when reloading from the sidebar (#1331)
The submission page would crash when reloading from the sidebar.

When reloading submissions, the page would crash because the rubric result was cleared from the store, causing the rubric viewer to `$emit('load', null)`, but the grade viewer did not ever expect it to emit `null`. The grade viewer now ignores the `load` event when no result is given.

Reloading courses/assignments would crash the page because suddenly the `assignment` prop on the page component was `null`, while we always expected it to be a valid `Assignment` model. Detecting on the submission page that the courses/assignments were reloaded was tricky, because we watched the assignment id, but we got that from the URL, which didn't change while the assignments were reloading. So now the logic of how we get an assignment/assignmentId on the submission page is flipped: we watch the URL to determine the `Assignment` model to be used, and derive the `assignmentId` from that.

This lead to another problem, however: the course store did not clear itself before reloading, so the assignment id on the submission page never actually changed, not even to `null` while we are reloading the assignments. So now the course store does that as well. Sadly, though, that broke a number of other pages, so they have been updated as well to account for the possibility of either `course` or `assignment` suddenly becoming `null`.

Anyway, now the `assignmentId` changes to `null` when we reload the assignments, and we can detect that we should reload the submission page's data.

Co-authored-by: Thomas Schaper <thomas@codegra.de>