Fix the newPage option
We should not have been unconditionally removing the last NewPage from
each subsection, we should have only done that on the last subsection.

This introduces a `mapCustom` function (in utils/typed.ts) that is
basically Array.map, but you can pass it additional functions to run for
the first or last element.