Improve plagiarism document rendering (#1335)
The LaTeX rendering function on the plagiarism page was rather unmaintainable,
and since we had a new feature request for the the output format we thought it
was time to rewrite it.

This implements a new document rendering library where you build an AST that
represents a document, which rendering backends can then render to their
specific format. The AST consists of a few basic nodes currently, but can be
easily extended if needed. We have implemented support for the following output
formats:

- LaTeX
- Docx

Right now we only use this to render the plagiarism report, but we may look
into other documents we want to render in the future.