Fix AutoTest results cache incoherence
On an `AutoTestRun` object we would keep two separate caches for AT
results: one for the latest results, and another mapping from user id to
the results of that user. We did not keep these two in sync properly,
though. Fixing this involved multiple separate fixes.

First, the `AutoTestResult` class had an `update` method, but also an
`updateExtended` method, which required one to know beforehand if the
update payload was regular or extended. While we know this in the places
where we make requests to the backend, this information is already lost
in the various `updateResults...` methods on the `AutoTestRun`. So we
always called `update` on the result there, even if we had extended
data.

`update` and `updateExtended` have been merged into a single method that
just checks if certain properties are available in the payload before
setting them.

Then there was an issue with the checks in the AT store's
`loadAutoTestResult` action that determine whether they should reload
a result if it is already present. It was possible that a result was
already present, but not in extended form, i.e. its step results weren't
loaded yet. We now check for the presence of step results, and if
they're not there we reload the extended result.

Finally there was the `resultsByUser` mapping on an `AutoTestRun`, which
mapped from user id to a list of results by that user. The idea was
initially to use this mapping to prevent us from showing the same user
multiple times in the results table, but we ended up not using it that
way because we do not have a list of user ids in the results table;
instead we ended up using it as an incredibly inefficient lookup table.

So this also gets rid of the `resultsByUser` mapping, and replaces it
with a `resultsById` mapping. Every time the results of the run are
updated, either because we have loaded the entire latest results list of
a run, or because we have loaded all results of a user, this mapping is
updated. The `results` list of an `AutoTestRun` now also always contains
the same objects as are present in the `resultsById` mapping, so when
a result is updated via its `update` method, this update will be
reflected in the `results` list.

This is not yet a proper fix to make all models immutable, that would
obviously require a lot more work, but I think it's at least an
improvement.