Remove "Add filter" button from analytics dashboard
The "Add filter" button on the analytics dashboard is a bit confusing
when splitting a filter, because then there are two primary buttons and
it's easy to mistake one for the other.

As discussed we would simply remove the button at the bottom of the
filters card, because we already have one at the top as well.