Basic support for IPython Notebook format v3 (#1346)
There were some breaking changes between IPython Notebook format v3 and v4, most notably the introduction of a data object on each output cell that contains the data we're interested in. In v3 the keys in that data object were simply set on the output cell itself, so now we try to find the data on the output cell itself if the data key is not present.

This also fixes the issue that we didn't check if code cells have an input prop instead of a source, something we did already account for in the case of a generic cell.

We now show an error when the notebook format version is less than 3, and a warning, along with the rendered notebook, if the version is equal to 3.

Finally, this adds an error handler on the <file-viewer> so that if any error is thrown from a subcomponent, the user will not be presented an empty file viewer.