Fix infinite loader when reloading old submission (#1345)
* Fix infinite loader when reloading old submission

For unknown reasons we removed the call to `loadCurrentSubmission` in
https://github.com/CodeGra-de/CodeGra.de/commit/ffd6782f0e5b797d8f8744d5c8dcd284cb03f7c0#diff-6375a0641e14f798e2596e2888f97651
which causes an infinite loader on the submission page because the
submission isn't ever loaded. This did not happen for the latest
submission, because the latest submissions of an assignment are always
loaded.

* Load the current submission when the assignment id changes

Reloading from the sidebar was still broken when on an old submission,
because `submissionId` watcher would never trigger with a valid
submission id when clicking the reload button in the sidebar, because it
would delete all submissions and only reload the latest submissions of
the assignment.