Fix errors on submissions page when assignment is hidden (#1347)
When an assignment is hidden no information about it is sent at all when the user does not have permission to see hidden assignments within the course. This caused some errors on the submissions page resulting in an infinite loader.

This adds some more checks to see if assignment is equal to null and shows a message to the user that the requested assignment does not exist or that the user does not have permission to see it yet.