DEV-183 Course feedback (#1342)
This adds a component that displays all feedback on all latest submissions in a course of a single user. This overview of feedback is available to teachers in the same location as the file tree on the submission page. Students have access to it via a new button on the submissions page.

This includes a new searching framework which allows you to specify the searchable properties in a list of objects within TypeScript's type system, and supports searching through all fields, or only a specific one.

This also makes the file tree resizable on small screens, i.e. when the file tree is rendered below the code viewer instead of to the right.

DEV-183 #done