Fix LTI 1.3 migration (#1366)
The problem was that we assumed that "users-courses.course_id" is a course
id (as the name suggests), however it is *not*. It is a course role id, and we
forgot to join the course role before going to the course. This commit fixes
that, and it also fixes the issue when a user was already connected to multiple
LTI 1.1 providers that it would only be connected to a single one after the
migration. This is fixed by simple getting all LTI courses that a user is
enrolled in and adding the lti provider ids for each of these courses (after
filtering out duplicate lti course ids of course).