Improve student buttons on submissions page
On a medium screen we would try to fit 3 buttons next to each other, but
this wasn't giving them enough space to render their labels on a single
line on screens just a bit larger than small screens. This makes the
buttons 40% of the available width, up to a minimum size that leaves
them enough space to draw their labels on a single line.