Fix AutoTest suite table rendering
On some screen sizes (slightly above `1200px` wide) the AutoTest suite
tables were drawn outside of the card sometimes.

At first I thought this happened because the `colspan` attributes on the
collapsed table rows were wrong. They were always set to 5, but when no
result is present the rest of the table has only 4 columns. This did not
fix the problem, though.

In the end I found that it is caused by `<code>` tags containing long
unbreakable words. I tried to address the problem in two ways:

* Increase the minimum screen size required to render the test suites in
  two columns by 25%. Before the split was at `1200px`, so this is now
  `1500px`. This means that a single-column suite can now become quite
  wide, but I think it's better than the alternative of two columns
  becoming very narrow on screens slightly wider than `1200px`.
* Set `word-break: break-word;` on the `<code>` tags in the step summary
  rows. This causes long words to break so that the table can always be
  rendered within the width it's given. It may not look very pretty or
  be very clear when words are broken, but with the other change above
  the tables will generally be wider than they were before, so I don't
  think that will happen a lot anymore.