Fix submission page category switching after initial load
The category switched because `defaultCat` was a computed property which
changed after deleting the final comment on a submission from the
feedback overview, when the feedback overview was the default category.

With this change the default category is set only once per submission,
so it will not change after the initial page load anymore.