DEV-123 Make it possible to run AutoTest on the teacher's revision (#1390)
This makes it possible to run an AutoTest on the teacher's revision instead of the student's submission. It adds a new option to an AutoTest, "Preferred revision" that must be set before the AT run can be started, similar to the "Grade calculation" and "Always visible" options.

This also changes all those options to use toggles that save immediately, instead of radios with a submit button.