Fix race condition in polling loop (#1399)
We only send the `approxWaitingBefore` only in the extend response, however on
the manage assignment page with a result open we have two polling loops: for all
the results (non extended) and for a single result. If the first polling loop
already sets a result to done we never get the final extended result. This means
that the `approxWaitingBefore` could still be an incorrect value, or more
importantly: we could miss the final step results.

Co-authored-by: Olmo Kramer <olmo@codegrade.com>