Fix calling loadData too frequently on Submissions page
`loadData` was called each time the `submissions` array changed, but
this was too often, as the computed `submissions` prop depends on the
user id and the assignment id. The former of which is not the problem,
but the assignment id is computed from the route which changes quite
often, especially on the analytics page. And somehow changing the route
triggered `assignmentId` and everything depending on it to be
recomputed, resulting in a short loader because `loadData` always sets
`loading` to `true`.

This is what actually caused the analytics integration tests to break.
Every time a value of a filter changed, the entire analytics dashboard
would be unrendered for a moment before appearing again, which caused
cypress aliases to reference stale DOM elements.

With this change we watch `userId` instead of `submissions`, which is
the only other dependency of `submissions` next to `assignmentId`, which
already has a watcher.