Proper fix for the analytics tests
The previous fix broke reloading submissions. This fixes that by
toggling `loadingInner` when reloading submissions, forcing a remount of
the components on the submissions page which gives them a chance to
perform a reload.

It also turned out that the analytics store was completely broken. It
stored its objects under a key `undefined`, always triggering a reload
of the data in the store. It also sometimes used only the workspace id
and sometimes a combination of assignment id and workspace id to index
the store. So this also rewrites part of the analytics store to function
correctly again.