Change information in the "Latest submission" button
We used to show the grade, but it is much more useful to know the latest
submission time. We haven't added more information such as the AutoTest
status, as we do not have that information yet on the submissions page,
and also the buttons may become too crowded. If you really want to see
more information about your submission, just go to the submission page.

DEV-197 #done