DEV-243 Use capture points step state icons for unit test step (#1403)
The new Unit Test step always had a green checkmark as its state, even when none of the tests passed. This changes its behaviour to be the same as the Capture Points test, i.e. a green checkmark only when all tests passed, a black tilde if some but not all tests passed, and a red cross when all tests failed or you achieved 0 points in some other way.

DEV-243 #finished