Improve the way the "Upgrade available" toasts is displayed
This changes two things:
1. It builds a static file that contains the commit hash of the latest front-end
version of CodeGrade available. This fixes the problem when the backend was
already restarted but the frontend was still building, in this case the toast
was visible but reloading didn't actually update the frontend. This also reduces
some of the load on the backend.
2. When the request fails for whatever reason we used to show the toast, now we
never show the toast as the most common case is that there is actually no new
version available.