Stop silently ignoring errors (#1404)
* Show a toast when changing the assignee fails

* Show a toast when changing course role of a user fails

This also fixes another issue with changing a user's role when an error
occurs: we used to update the user whose role had changed before sending
the request, without reverting the changes if the request failed. Now we
only update the user's role when the request was successful.

This also fixes an infinite loader in the roles dropdown when an error
occurred.