Add patchAssignment action on courses store
This replaces a lot of other actions in the courses store with one that
takes all assignment data returned from the server and stores a new
Assignment model, rather than updating the one that is currently in the
store, which prevents data from becoming out of sync. For example,
changing the "available at" property of an assignment might also change
its state, but we did not account for that in
updateAssignmentGeneralSettings.